+++ doc/releases/release_notes.rst
@@ -21,6 +21,10 @@ Developpement version
 - ``part_tree_to_file``: manage user provided links, and write CGNSBase_t misc. children
 - ``interpolate``: managed Vertex located fields for all values of ``strategy`` parameter
 
+🐞 Fixes
+^^^^^^^^
+- ``transform_affine``: update periodic values of GridConnectivity_t nodes
+
 🚧 API change
 ^^^^^^^^^^^^^
 - ``part_tree_to_file``: add ``links`` parameter